LAST FIXES
