--- lib/features/car_details/presentation/widgets/car_details_text_field.dart
@@ -1,25 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CarDetailsTextField extends StatelessWidget {
-  const CarDetailsTextField({
-    super.key,
-    required this.labelText,
-    required this.hintText,
-  });
-
-  final String labelText;
-  final String hintText;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextField(
-      decoration: InputDecoration(
-        labelText: labelText,
-        hintText: hintText,
-        border: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(15),
-        ),
-      ),
-    );
-  }
-}
